third_party/blink/web_tests/external/wpt/editing/run/inserttext_1-1000-expected.txt

This is a testharness.js-based test.
Found 46 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["inserttext","\\t"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo\\tbar" but got "foo<span style=\\"white-space:pre\\">\\t</span>bar"
[FAIL] [["defaultparagraphseparator","div"],["inserttext","\\n"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo</div><div>bar</div>" but got "foo<div>bar</div>"
[FAIL] [["defaultparagraphseparator","p"],["inserttext","\\n"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<p>bar</p>"
[FAIL] [["defaultparagraphseparator","div"],["inserttext","abc\\ndef"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>fooabc</div><div>defbar</div>" but got "fooabc<div>defbar</div>"
[FAIL] [["defaultparagraphseparator","p"],["inserttext","abc\\ndef"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>fooabc</p><p>defbar</p>" but got "fooabc<p>defbar</p>"
[FAIL] [["inserttext"," "]] "foo &nbsp;[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo []&nbsp;bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo[] &nbsp;bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo&nbsp; []bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo&nbsp;[] bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo[]&nbsp; bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo&nbsp;&nbsp;[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo&nbsp;[]&nbsp;bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo[]&nbsp;&nbsp;bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; bar" but got "foo&nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo []&nbsp;        bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp;&nbsp; bar" but got "foo&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo []&nbsp;&nbsp; &nbsp; bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp; &nbsp; &nbsp; bar" but got "foo&nbsp; &nbsp; &nbsp; &nbsp;bar"
[FAIL] [["inserttext"," "]] "foo&nbsp;[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp;" but got "foo&nbsp;&nbsp;"
[FAIL] [["inserttext"," "]] "foo&nbsp;{}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp;" but got "foo&nbsp;&nbsp;"
[FAIL] [["inserttext"," "]] "foo&nbsp;&nbsp;[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp;&nbsp;" but got "foo&nbsp; &nbsp;"
[FAIL] [["inserttext"," "]] "foo&nbsp;&nbsp;{}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo &nbsp;&nbsp;" but got "foo&nbsp; &nbsp;"
[FAIL] [["inserttext"," "]] "<b>foo[]</b>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foo </b>bar" but got "<b>foo&nbsp;</b>bar"
[FAIL] [["inserttext"," "]] "foo[]<b>bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo <b>bar</b>" but got "foo&nbsp;<b>bar</b>"
[FAIL] [["inserttext"," "]] "foo[]<span> </span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo&nbsp;<span> </span>" but got "foo&nbsp;<span></span>"
[FAIL] [["inserttext"," "]] "foo[]<span> </span> " compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo&nbsp;<span> </span> " but got "foo&nbsp;<span></span>"
[FAIL] [["inserttext"," "]] "<span> </span>[]foo" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span> </span>&nbsp;foo" but got "<span></span>&nbsp;foo"
[FAIL] [["inserttext"," "]] " <span> </span>[]foo" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected " <span> </span>&nbsp;foo" but got " <span></span>&nbsp;foo"
[FAIL] [["inserttext"," "]] "<p>foo&nbsp;[]<p>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo &nbsp;</p><p>bar</p>" but got "<p>foo&nbsp;&nbsp;</p><p>bar</p>"
[FAIL] [["inserttext"," "]] "<div style=white-space:pre-line>foo &nbsp;[]bar</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"white-space:pre-line\\">foo &nbsp; bar</div>" but got "<div style=\\"white-space:pre-line\\">foo&nbsp; &nbsp;bar</div>"
[FAIL] [["inserttext"," "]] "<div style=white-space:pre-line>foo&nbsp;[]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"white-space:pre-line\\">foo &nbsp;</div>" but got "<div style=\\"white-space:pre-line\\">foo&nbsp;&nbsp;</div>"
[FAIL] [["inserttext"," "]] "<div style=white-space:nowrap>foo &nbsp;[]bar</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"white-space:nowrap\\">foo &nbsp; bar</div>" but got "<div style=\\"white-space:nowrap\\">foo&nbsp; &nbsp;bar</div>"
[FAIL] [["inserttext"," "]] "<div style=white-space:nowrap>foo&nbsp;[]</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"white-space:nowrap\\">foo &nbsp;</div>" but got "<div style=\\"white-space:nowrap\\">foo&nbsp;&nbsp;</div>"
[FAIL] [["inserttext"," "]] "http://a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://a\\">http://a</a>&nbsp;" but got "http://a&nbsp;"
[FAIL] [["inserttext"," "]] "ftp://a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"ftp://a\\">ftp://a</a>&nbsp;" but got "ftp://a&nbsp;"
[FAIL] [["inserttext"," "]] "quasit://a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"quasit://a\\">quasit://a</a>&nbsp;" but got "quasit://a&nbsp;"
[FAIL] [["inserttext"," "]] ".x-++-.://a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected ".<a href=\\"x-++-.://a\\">x-++-.://a</a>&nbsp;" but got ".x-++-.://a&nbsp;"
[FAIL] [["inserttext"," "]] "(http://a)[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "(<a href=\\"http://a\\">http://a</a>)&nbsp;" but got "(http://a)&nbsp;"
[FAIL] [["inserttext"," "]] "&lt;http://a>[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "&lt;<a href=\\"http://a\\">http://a</a>&gt;&nbsp;" but got "&lt;http://a&gt;&nbsp;"
[FAIL] [["inserttext"," "]] "http://a![]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://a\\">http://a</a>!&nbsp;" but got "http://a!&nbsp;"
[FAIL] [["inserttext"," "]] "!\\"#$%&amp;'()*+,-./:;&lt;=>?^_`|~http://a!\\"#$%&amp;'()*+,-./:;&lt;=>?^_`|~[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "!\\"#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~<a href=\\"http://a!&quot;#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~\\">http://a!\\"#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~</a>&nbsp;" but got "!\\"#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~http://a!\\"#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~&nbsp;"
[FAIL] [["inserttext"," "]] "http://a!\\"'(),-.:;&lt;>`[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://a\\">http://a</a>!\\"'(),-.:;&lt;&gt;`&nbsp;" but got "http://a!\\"'(),-.:;&lt;&gt;`&nbsp;"
[FAIL] [["inserttext"," "]] "http://a#$%&amp;*+/=?^_|~[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://a#$%&amp;*+/=?^_|~\\">http://a#$%&amp;*+/=?^_|~</a>&nbsp;" but got "http://a#$%&amp;*+/=?^_|~&nbsp;"
[FAIL] [["inserttext"," "]] "mailto:a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"mailto:a\\">mailto:a</a>&nbsp;" but got "mailto:a&nbsp;"
[FAIL] [["inserttext"," "]] "a@b[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"mailto:a@b\\">a@b</a>&nbsp;" but got "a@b&nbsp;"
[FAIL] [["inserttext"," "]] "#@x[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"mailto:#@x\\">#@x</a>&nbsp;" but got "#@x&nbsp;"
[FAIL] [["inserttext"," "]] "!\\"#$%&amp;'()*+,-./:;&lt;=>?^_`|~a@b!\\"#$%&amp;'()*+,-./:;&lt;=>?^_`|~[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "!\\"#$%&amp;'()*+,-./:;&lt;=&gt;<a href=\\"mailto:?^_`|~a@b\\">?^_`|~a@b</a>!\\"#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~&nbsp;" but got "!\\"#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~a@b!\\"#$%&amp;'()*+,-./:;&lt;=&gt;?^_`|~&nbsp;"
[FAIL] [["inserttext"," "]] "<b>a@b</b>{}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"mailto:a@b\\"><b>a@b</b></a> " but got "<b>a@b&nbsp;</b>"
Harness: the test ran to completion.
